Made a seperate FXML file and controller class for each tab.
