+++ src/main/java/controllers/SimulatorController.java
@@ -1,0 +1,4 @@
+package controllers;
+
+public class SimulatorController {
+}
